renamed slice to api
