+++ src/redux/stores/createRootReducer.ts
@@ -2,12 +2,12 @@ import { combineReducers } from '@reduxjs/toolkit';
 
 import { History } from 'history';
 import { connectRouter } from 'connected-react-router';
-import { authApi } from './auth/authSlice';
+import { authApi } from './auth/authApi';
 import wsReducer, { wsReducerPath } from './ws/wsSlice';
 import chatReducer, { chatReducerPath } from './chats/chatSlice';
 import errorReducer, { errorReducerPath } from './apiErrors/errorSlice';
-import { userApi } from './user/userSlice';
-import { profileApi } from './userProfile/userProfileSlice';
+import { userApi } from './user/userApi';
+import { profileApi } from './userProfile/userProfileApi';
 
 const createRootReducer = (history: History) =>
   combineReducers({
